Make statsgrid use handlebars helpers
Stats now use handlebars helpers, and the entries in template for
movement allowance and body type have been, for consistency's sake,
renamed to "ma" and "bt". To allow for consistent translation, then are
"Ma" and "Bt" in the translation file, as stats use proper case. This
will likely be a breaking change for existing characters.