Initial addition of derived stats row
Derived stats like BTM, carry and such. There wasn't room for them to
the side of the regular stats I think, so I added them below. The
spacing is important - they should be close to the main stats, but a
little far from the damage tracker to imply they're not related so.