Add chip icons for skills
This... was a surprising bastard to get any kind of right, and may be a
bit of a spaghetti mess. Chip icon has to be inside a flex div; otherwise,
when the element expands vertically, there's just a bunch of empty space
under it.

Oh, and there's no licensing worry with the icon. I made it, terribly,
pretty fast. Anyone forking the system has free permission to use it.