Initial skills sorting
It's not accessible by the sheet yet - you have to set the actor's
skillSortOrder flag, then set the sortSkills flag to trigger a skills
sort on the next time actor data is processed. eg

actor.setFlag("cyberpunk2020", "skillSortOrder", "alph")
actor.setFlag("cyberpunk2020", "sortSkills", true)

skillSortOrder is by default "alph" if not present, and may be set to
"alph" or "stat".

The reason I haven't done it in the sheet data/sheet processing is cos
sorting 89 stats (at minimum) every time a sheet opens is a little cheeky.
This may have to change, as I intend players to be able to search
skills, and I can't just remove skills in the main skills list.