Finished field refactor
Whew. This one has a few changes bundled in. I'm sorry about so much
being in one commit! It was a whole lot of changes to make things work,
and I made a few others on the way.

Field is now pretty much in its own little scss file, and makes for VERY
convenient styling of anything with a label and an input, as well as
a small image.

The armor display is a little more complete too. I'd actually tried a
table layout that would allow for displaying armor segments (eg torso)
in rows, with total stopping power, hit location numbers, as well as
possibly splitting hard and soft armor. It was quite condensed, and
would have allowed easy expansion to view the armor that actually
covered that segment.
After showing them both to a couple of my players though, they'd prefer
the armor layout in this comment, with a click bringing up a dialog with
that armor.

That dialog isn't in, but the armor display is.