A start on work for selecting attack modifiers.
The idea is to display a window in which players can select which
modifiers are currently active, which can then be used for weapon
rolling.

It's worth noting that at the moment, this is a very rigid
implementation start; I'd like to see something a little more flexible.
In any case, it's my first time using FormApplication and I'm
essentially nabbing form the 5e system.