First working modifiers!
Hoo boy that was a bit of a journey.
In a nutshell, clicking a weapon in the combat tab (and this should be
changed in future to clicking an Attack button that shows on hover), a
dialog shows: this is AttackModifiers, a form application. It shows a
bunch of modifiers that can be had for firing a weapon, allowing you to
select firing mode, range, and a few other things.

It *is* a little rigid. I'm not sure I'm happy with a class being
dedicated to just attack modifiers, with such a static setup either.
This is especially true with the fact that melee isn't even involved
here yet.

What does still need to be done:
* target selection (depends on firemode, and might be a reason to put
firemode in a previous dialog)
    * full auto. Give targets/number of, and split RoF between them. At
    the moment full auto dumps into one target.
    * 3rb. One target, but should be selectable.
    * Suppressive. Area. Ugh.
    * Other area of effect weapons such as claymore, flamethrower,
    autoshotguns.
* select number of bullets to fire
    * likely limited by bullets left in gun and rate of fire. This is
    actually calculated in the full auto mods calcing code, but there's
    no way for players for edit this.
* Do changes for shooting multiple rounds
    * full auto card that splits hits by locations, orders by hit roll,
    and totals damage
    * apply same methodology to 3rb?
* aim at body locations
    * this'd likely be in the modifiers box. Choose body part, tada.
    Easy enough.

* melee. hoo boy melee. we'll get to that