Wounds reset bug fixed
Weird rabbithole here - the actor._data didn't
reflect the change to damage, which meant that data.damage got
overridden in prepareData(). I'd thought I was setting it wrong
somewhere, but the ACTUAL bug was that no wound radio button was
actually selected by default, meaning that if none got clicked, the
value of _data.damage would be assigned "null".

damage in damageBoxes is actually a string, and I'm not entirely sure
how to get it consistently being a number, so I added an isChecked to
damageBoxes that uses == to check which box should be checked instead of ===.