Fix wounds toggling bug
The intended behaviour of the wound tracker is that if the highest
filled wound on the wound tracker is clicked, that wound will be
removed. With that wound removed, clicking in the same place would
reinstate that wound. This allows for toggling being uninjured without
having a specific "uninjured" button.

This behaviour existed somewhat beforehand, but broke whenever anything
about the actor changed. The checked attribute was added to the wound
radio buttons to ensure data persisted, which should have been done in
the first place. That fixed losing wound data, but it introduced the bug
that toggling the highest wound no longer worked.
What broke it is radio buttons' intended behaviour. It will always have
ONE radio selected. To keep the damage of the actor intact, the radio
selected must match the damage value the actor is currently at - which
means it *cannot* change value if we click it. Checkboxes don't quite
work either, as there is only one "damage" value at any given time.

So, I made the wound boxes not a radio anymore. They are now spans with
click events to update the actor's damage, and this works quite well,
though it loses a slight amount of semantic meaning.