Sorting reverted back to "I need this" flag
By default was trying to trigger an entity update before the entity was
even in the list of entities. So I've ordered the skills in
template.json to allow for them to START ordered alphabetically.

Thing is, ordering by stat straight-up does not work at the moment, will
have to investigate this. Likely to do with Object.entries not caring
how the object is defined.