Fix sort bug mostly, add migration
Okay, so the sort bug was mostly a bug because there was an
inconsistency in what stats it ordered the stats by, as well as it not
having a rank for EACH stat. It was using "body" instead of "bt", and
only had ranks for a few of the stats.

That led me to finding that the skills that use body type (skills.bt)
actually had stat "body". Ahh. That's not cool. So, change those to "bt"
in the template json, and fill out the stat rankings in skill-sort
properly.
Except we're kind of at the point that people are starting to use the
system, or will be, and this makes a change to skills. So, to not break
their stuff, I have to migrate it. I mostly nabbed the framework for
this from dnd5e's migration. The migration from skills' body to bt DOES
work, but I need to actually make sure that this migration is applied
for someone moving from 0.1.2-alpha to 0.1.3, or any pre-migration
version to migration. Need to test that.