Sorting skills should Actually Work
Okay. So sorting was a NIGHTMARE with how I had things previously.
Skills are a big object in actor.data.data - an object so that rolls can
go @skills.bt.total etc. Except sorting needs ORDER. And I was trying to
do it in-place. And relying on JS objects for order is...rocky, even recently. Don't do it.
Because of this, to get it to update, I had to get foundry to empty the
object before replacing it with a sorted object, lest it think they're
the same and not even change it. This made for some NASTY loops that
were hiding bugs I didn't even want to delve into - code smells were
popping up.

So, entirely different approach. Just store a list of skill names that
determine what gets displayed. Ask an actor to sort? It'll note how you
sorted it (so the select on the sheet can show that), and store the
newly-sorted display order. Tada. Done. Things don't crash and burn. And
it's a lot less painful.