Weapons select attack skill
Pistols: Handgun
SMGs: SMG
Melee stuff: Fencing/melee
Rifle: Rifle
Shotgun: Rifle
Heavy: Heavy Weapons
Exotic: Literally any skill.

The not-so-great part at the moment is that when the weapon type
changes, the skill used should immediately follow suit. It doesn't - if
I set the skill used in getData(), it's not permanent. If I call
update(), I think it's too late. It's not so much of an issue right now,
as when the sheet closes it is updated correctly, but it still rubs me
the wrong way.