Change field-list from flex to grid
The big reason for this is that it was *really hard* to have gaps
between fields in the list but not at the edge; margin-right would make
for an extra, weird gap on the right of the container, so it wouldn't
stretch to the edge.

The "gap" property solves this neatly. It is gorgeous. And even applies
to flexbox... but only for chromium 84. Foundry's Electron as of 0.7.9
has chrome 80, so we don't have access to that. So we have to use grid
to access it.

Grid is honestly about as flexible as flexbox here; to adjust amount of
columns, instead of adjusting children's max-width, you adjust
grid-template columns to have another space. Neato.

Anyways, there should be no more of those weird margins around!