Top-level files and declarations renamed to cyberpunk2020
They were previously just cyberpunk. Thing is, CPRED is a thing too, so
I think it's a good idea to, at least at the top level, refer to
specifically 2020. The system itself, and the repo, are now named as
such. The classes etc within those files are still just "cyberpunk" - by
this point we've made the distinction of WHICH cyberpunk it is, and are
in this one. That, and "cyberpunk2020" is unwieldy for class names, and
"cp2020" is even more terse than "cpRED".