Actors and token tweaks
There was a recursion bug when opening token settings for actors with
items. By storing equipped items in data, we gave a route for recursion
to visit the actor's data infinitely. So it's not done anymore!

NPCs are now added - they are functionally the same as actors, but their
tokens do not get vision by default, and are not linked to their actors,
whereas characters are. There is an indication on their sheet, next to
their name, saying they are NPCs.

All character actors' tokens should now have vision and be linked to
their token.