Make modifiers infinitely more generic
Okay, so previously, attack-modifiers was hardcoded for ranged
modifiers. This was kinda fine as I hadn't approached other modifiers
yet, and was serving as a kind of proof-of-concept. It's stopped being
fine, cos I need to do the melee stuff.

So to make this more generic, I needed to make a way of describing
modifiers (an example of this can be seen as rangedModifiers in
`lookups.js`). Most of this is easy - just give a localisation key,
localisation data if needed as well, a data path to say where in the
resulting object this modifier will be stored, and a default value so
the field type can be inferred. (data path is necessary if we want to
describe modifiers in arrays, which we do because we want to be able to
control the order it's presented)

Where it gets harder is anything with choices. Because each choice can
have a separate localisation data, value, localisation key etc. For
choices so far, I'd been storing a choices `["Just", "Choice",
"Values"]`, a locals `["What", "User", "Sees"]`, and localData
`[{"Just":"foo"}]` used if there needed to be anything transplanted in
localisation. But for modifiers, I'd been storing them per modifier, so
instead of `[choices, locals, localData]` as `[{localKey, localData,
value}, ...]`. This made a lot more sense for when you're describing
modifiers, but the two were inconsistent, I didn't want BOTH ways of
doing it in there. The second was better for describing modifiers, the
first for letting select boxes just be given strings for choices.

I ended up picking the best of both worlds. Selects now use a handlebar
helper that essentially translates a bare "choice text" array into the
more complex version when needed. Whew

But yes, now modifiers can be supplied at runtime to ModifiersDialog,
and in groups so that there's a mild degree of control over display.
Please see `lookups.js -> rangedModifiers` for a good example, and
`actor-sheet.js -> activateListeners` for how those modifiers are given
to the class.

Hopefully this should let me easily supply custom modifiers everywhere!
By the way, it will NOT fully work when you hit OK - that bit is still
hardcoded. It's an easy change, but it's 20 past midnight for me, so
calling it a night.
Sorting is also possible.