Add roll modifiers to attack in autofire
Okay admittedly this doesn't yet look as good as the regular rolls, but
we'll get there - I want to be able to, given just a regular roll, style
it up and spit it out. *preferably* long form and a mouseover form, so
it can be done for the damage table too.