Go to the sane sheet data model
In 0.8.x, in handlebars you'd do {{data.data.foo}}, but for field names
"name=foo". This feels mildly insane, breaks a few things, and
unbreaking it by making some things data.data feels difficult to remain
consistent with or make sense of.

So this one just applies actor's actor.data.data in the `data` field of
the sheet data, as it was before. Now we are once more at {{data.foo}}
instead of {{data.data.foo}}