Supporting melee attacks differently
They're now not the same as ranged attacks! And it's quite close to
having martial arts in there :) There's still an issue that in attack
skills of a weapon, martial arts won't translate, but I'll sort that at
some point.

Select fields now have the ability to take option groups! give it a
choices of, for example:
[{groupName: "Group name!", choices: ["one", "two", "three"]},
{groupName: "Other group", choices: ["aaa", "bbb"]}]