Remove the concept of translation prefixes
When something in the localization file was SkillPREFIXMartial for
example. Ugh, it only caused headaches to check where that was involved.
Besides - if you want "Martial Arts" in a text, doesn't matter what
context it's in, it's *still the same pair of words*.

Anyways, this should simplify translation of subskills quite nicely.