Quick and dirty method to make a compendium from template skills
Essentially this should produce a compendium with all vanilla skills in.
It also reflects a couple tentative decisions - I'm not convinced I
should have skills that can contain other skills, here. Mayyybe an item
type that's a skill group, but then I have to do a sheet for that too,
and it'd be hard to access those skills individually... ugh. In any
case, for now this generates a flat list of skills, with nested ones
looking like "Pilot: Gyro" etc.