Fix automatic migration not working
So, I recently figured out I'd been doing a goof, storing the last
system migration setting with `game.settings.set("cyberpunk",
"systemMigrationVersion")`. That's been fixed recently, and is now under
"cyberpunk2020", and correctly recognised as one of the system's
settings.
Thing is, it was set to default at the game's current version. Seeing as
migration only occurs if `NEEDS_MIGRATION_VERSION` is newer than the
last time migration occurred... well, it basically never triggered. So,
it's now set up to be empty by default, and that'll mean there'll be a
migration the first time the system is ever run. It should also mean
that even if you open a world from before this update, that world gets
migrated.