Right-click to delete items in gear or skills
A confirm dialog will pop up, which now tells you which item you're
trying to delete. This is a somewhat stopgap measure, as deleting this
way isn't particularly intuitive - some users won't realise its even
possible.

But for now, it stops things moving about when mousing over, and it also
allows for not cluttering up the UI everywhere (if I were to have the
delete button show all the time)

It may be worth doing a slightly transparent delete button for items
that uses the font-awesome delete bin; it'd have to be both small and
unobtrusive though