Bandaid fix for martial arts name translation
So, until there's a full martial arts update which adds an actual field
for martial arts skills to indicate they're indeed martial arts skills,
we've thus far indicated they're martial arts by having their *actual
name* be "Martial Arts: ____" in the default skill compendium that
populates new characters. Migration code also does this. It lets even
pre-martial-update code check if something's martial arts, and allows
the user to just search "Martial" on the default populated skills.

As we have "Martial Arts: ____" for now, we can allow such martial arts
to be translated by changing the translation keys to match (and
thankfully, these still work with spaces etc in).

When (if? this is pretty on-backburner) martial arts gets implemented
more fully, we can hopefully try to transition to JUST having "Judo",
"Boxing" etc for the names, and a checkbox for if the skill is Martial.
I think some work for that is already done in the `develop branch`.
Which past-me should have probably started in a dedicated branch, but
hey-ho I've sure as heck learned.

I'd have liked the actual names of default skills to be their
translation keys (eg "KickBoxing") and those get translated in the
skill-list, but that gets complex when it comes to intermixing with
custom-created skills, as well as detecting whether someone is trained
in specific default skills (for example, Solo checking if someone's
trained in Awareness. What about Awareness in Spanish?).
I haven't yet tackled that (and it's apparently a
KNOTTY issue), so for now the skills have their Full And Concrete
Non-Translating Names as their actual names.