Fix new actors not having skills, or being able to see
I think the previous method f doing it got dprecated, as I was told in
no uncertain terms that the objects I was manipulating are not
manipulable. So w go with the old trusty updates.

I wouldn't be surprised if token vision was a little diffrent than
previous though; I remember seeing something about this changing.