Update system.json compatibility
From some brief testing, the system seems to work on Foundry v11.
Changed `system.json` to reflect this. There also seems to have been a
slight change in syntax (or at least addition of capabilities) of the
file that let you specify version compatibility more specifically. I've
made use of this to keep the minimum at 10 but change verified to 11.

I've also added cybersmily and drowzero0 as authors, because they've
added quite a bit to the system.