Allow an actor's sheet to roll initiative on tracker
FoundryVTT has an encounter tracker in which tokens roll initiative.
You can successfully add a token to this by right-clicking it, then
clicking the swords and shield icon that pops up - from there you can
roll initiative. Nice.

Except, sheets often have a convenience button to roll initiative. The
button this system had rolled the initiative formula, but didn't add it to
the tracker. So today I went to go fix that.

The first place I went to was the Actor API, within which there was a
function rollInitiative(). *excellent*, I'll use that. It turned out I
was accidentally shadowing it, but... damn, it still won't show up in
the tracker.
Okay, there's an option to create combatants, that makes the actor
actually show up there... but it still doesn't roll, and still doesn't
with the option "rerollInitiative" set to true. There are no further
documented options.
So, it appears the only option left to me is to add the actor to the
combat with rollInitiative(), and THEN say to the combat "hey, this
actor? can you roll their initiative please?"

Anyways. This seems to work.