Fix #141
Skills weren't being copied when actors were. This is because actors, on
their creation, fetch the default skills and add them all to their
inventory.

This is, I agree, a massive pain. In most systems, this wouldn't be
necessary, but in CP2020 you can explicitly just add arbitrary skills.
Foundry - reasonably - expects a single shape for your Actor that
doesn't morph all the time, as defined in `template.json`
(would anything outside that shape be disregarded?)
This means to support that, I have the skills as items. These are in the
cyberpunk2020.default-skills compendium.

It is DEFINITELY worth experimenting with "can I get these to be part of the actor
instead?" as it means with skills specifically I do hacks to update them
when the Actor form is edited.

Anyways. Actors now check to see if they have any skills existing
already. If they do, they leave their skills alone.